+++ src/godok/dto/godok-status.dto.ts
